refactor : 데이터 호출 에러 처리 삭제 - 전역 처리로 필요 없어짐
