+++ apps/admin/hooks/useContent.ts
@@ -8,13 +8,9 @@ export const useContent = (testId: string) => {
   const [contentData, setContentData] = useRecoilState<MbtiTest>(mbtiTestDataState);
 
   const getContent = async () => {
-    try {
-      const response = await getContentAPI(testId);
-      if (response) {
-        setContentData(response.data.test);
-      }
-    } catch (error) {
-      alert(`error: ${error}`);
+    const response = await getContentAPI(testId);
+    if (response) {
+      setContentData(response.data.test);
     }
   };
